fix(FOROME-660): fixed inh mode
