+++ src/pages/filter/ui/panels/function-panel/function-panel.store.ts
@@ -8,7 +8,10 @@ import {
   TVariant,
 } from '@service-providers/common/common.interface'
 import { getQueryBuilder } from '@utils/getQueryBuilder'
-import { TScenario } from './function-panel.interface'
+import {
+  IInheritanceModeCachedValues,
+  TScenario,
+} from './function-panel.interface'
 
 class FunctionPanelStore {
   constructor() {
@@ -23,6 +26,10 @@ class FunctionPanelStore {
   public get complexVariants(): [string, number][] {
     const { variants } = filterStore.statFuncData
     return variants || []
+  }
+
+  public get filteredComplexVariants(): [string, number][] {
+    return this.complexVariants.filter(([, variantValue]) => variantValue > 0)
   }
 
   public get filterName(): string {
@@ -115,6 +122,34 @@ class FunctionPanelStore {
       }
     }
     return '--'
+  }
+
+  public handleSelectAllVariants(problemGroupValues: string[]): void {
+    if (this.filteredComplexVariants.length === 0) return
+
+    const variantsGroup = this.filteredComplexVariants.map(
+      variant => variant[0],
+    )
+
+    this.setCachedValues<IInheritanceModeCachedValues>(
+      FuncStepTypesEnum.InheritanceMode,
+      {
+        conditions: { problem_group: problemGroupValues },
+        variants: variantsGroup,
+      },
+    )
+  }
+
+  public handleResetAllFieldsLocally(problemGroupValues: string[]): void {
+    if (problemGroupValues.length === 0) return
+
+    this.clearCachedValues(FuncStepTypesEnum.InheritanceMode)
+  }
+
+  public handleResetVariantsLocally(variantsValues: string[]): void {
+    if (variantsValues.length === 0) return
+
+    this.clearCachedValues(FuncStepTypesEnum.InheritanceMode, 'variants')
   }
 }
